+++ oh.sh
@@ -386,8 +386,13 @@ fi
 }
 
 function config_database {
+	echo "Checking if MySQL is running..."
+	if [ -f "$OH_PATH/$TMP_DIR/mysql.sock" ] || [ -f "$OH_PATH/$TMP_DIR/mysql.pid" ] ; then
+		echo "MySQL alreay running ! Exiting."
+		exit 1
+	fi
+
 	echo "Checking for MySQL config file..."
-	
 	if [ $GENERATE_CONFIG_FILES = "on" ] || [ ! -f ./$CONF_DIR/my.cnf ]; then
 		[ -f ./$CONF_DIR/my.cnf ] && mv -f ./$CONF_DIR/my.cnf ./$CONF_DIR/my.cnf.old
 
@@ -911,7 +916,8 @@ fi
 # if SERVER mode is selected, wait for CTRL-C input to exit
 if [ $OH_MODE = "SERVER" ]; then
 	echo "Open Hospital - SERVER mode started"
-	echo "Database server listening on $DATABASE_SERVER:$DATABASE_PORT"
+#	echo "Database server listening on $DATABASE_SERVER:$DATABASE_PORT"
+	echo "Database server ready for connections..."
 	echo "Press Ctrl + C to exit"
 	while true; do
 		trap ctrl_c INT
